@@ -20,78 +20,78 @@
     "i18n": "tsc tools/i18n-cli.ts --lib dom,es2017 --target es5 --outDir tmp && node tmp/tools/i18n-cli.js"
   },
   "dependencies": {
-    "classnames": "2.2.5",
-    "clipboard": "2.0.1",
-    "codemirror": "5.38.0",
-    "color": "3.0.0",
-    "core-js": "2.5.6",
-    "file-saver": "1.3.8",
+    "classnames": "2.2.6",
+    "clipboard": "2.0.4",
+    "codemirror": "5.42.2",
+    "color": "3.1.0",
+    "core-js": "2.6.2",
+    "file-saver": "2.0.0",
     "golden-layout": "1.5.9",
-    "inversify": "4.13.0",
+    "inversify": "5.0.1",
     "inversify-inject-decorators": "3.1.0",
     "isomorphic-fetch": "2.2.1",
     "jed": "1.1.1",
     "jquery": "3.3.1",
     "keymaster": "1.6.2",
-    "markdown-it": "8.4.1",
+    "markdown-it": "8.4.2",
     "node-noop": "1.0.0",
     "po2json": "0.4.5",
-    "react": "16.4.0",
-    "react-dom": "16.4.0",
-    "react-router": "4.2.0",
-    "react-router-dom": "4.2.2",
+    "react": "16.7.0",
+    "react-dom": "16.7.0",
+    "react-router": "4.3.1",
+    "react-router-dom": "4.3.1",
     "reflect-metadata": "0.1.12",
-    "rxjs": "5.5.10",
+    "rxjs": "6.3.3",
     "spark-md5": "3.0.0",
-    "tslib": "1.9.1"
+    "tslib": "1.9.3"
   },
   "devDependencies": {
-    "@types/chai": "4.1.3",
+    "@types/chai": "4.1.7",
     "@types/chai-as-promised": "7.1.0",
-    "@types/classnames": "2.2.3",
-    "@types/clipboard": "2.0.0",
-    "@types/codemirror": "0.0.56",
+    "@types/classnames": "2.2.7",
+    "@types/clipboard": "2.0.1",
+    "@types/codemirror": "0.0.71",
     "@types/color": "3.0.0",
-    "@types/core-js": "0.9.46",
-    "@types/enzyme": "3.1.10",
-    "@types/enzyme-adapter-react-16": "1.0.2",
+    "@types/core-js": "2.5.0",
+    "@types/enzyme": "3.1.15",
+    "@types/enzyme-adapter-react-16": "1.0.3",
     "@types/esprima": "4.0.2",
-    "@types/file-saver": "1.3.0",
-    "@types/glob": "5.0.35",
-    "@types/history": "4.6.2",
+    "@types/file-saver": "2.0.0",
+    "@types/glob": "7.1.1",
+    "@types/history": "4.7.2",
     "@types/isomorphic-fetch": "0.0.34",
-    "@types/jquery": "3.3.2",
+    "@types/jquery": "3.3.29",
     "@types/keymaster": "1.6.28",
-    "@types/markdown-it": "0.0.4",
-    "@types/mocha": "5.2.0",
-    "@types/node": "10.1.2",
-    "@types/react": "16.3.14",
-    "@types/react-dom": "16.0.5",
-    "@types/react-router": "4.0.25",
-    "@types/react-router-dom": "4.2.6",
-    "@types/react-test-renderer": "16.0.1",
-    "archiver": "2.1.1",
-    "chai": "4.1.2",
+    "@types/markdown-it": "0.0.7",
+    "@types/mocha": "5.2.5",
+    "@types/node": "10.12.18",
+    "@types/react": "16.7.18",
+    "@types/react-dom": "16.0.11",
+    "@types/react-router": "4.4.3",
+    "@types/react-router-dom": "4.3.1",
+    "@types/react-test-renderer": "16.0.3",
+    "archiver": "3.0.0",
+    "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
-    "clean-webpack-plugin": "0.1.19",
-    "copy-webpack-plugin": "4.5.1",
-    "css-loader": "0.28.11",
-    "enzyme": "3.3.0",
-    "enzyme-adapter-react-16": "1.1.1",
-    "esprima": "4.0.0",
-    "file-loader": "1.1.11",
-    "fork-ts-checker-webpack-plugin": "0.4.1",
+    "clean-webpack-plugin": "1.0.0",
+    "copy-webpack-plugin": "4.6.0",
+    "css-loader": "2.1.0",
+    "enzyme": "3.8.0",
+    "enzyme-adapter-react-16": "1.7.1",
+    "esprima": "4.0.1",
+    "file-loader": "3.0.1",
+    "fork-ts-checker-webpack-plugin": "0.5.2",
     "git-revision-webpack-plugin": "3.0.3",
-    "glob": "7.1.2",
-    "handlebars": "4.0.11",
-    "handlebars-loader": "1.7.0",
+    "glob": "7.1.3",
+    "handlebars": "4.0.12",
+    "handlebars-loader": "1.7.1",
     "html-webpack-plugin": "3.2.0",
-    "husky": "0.14.3",
+    "husky": "1.3.1",
     "istanbul": "0.4.5",
     "istanbul-instrumenter-loader": "3.0.1",
-    "istanbul-lib-instrument": "1.10.1",
+    "istanbul-lib-instrument": "3.0.0",
     "json-loader": "0.5.7",
-    "karma": "2.0.2",
+    "karma": "3.1.4",
     "karma-bamboo-reporter": "0.1.2",
     "karma-chrome-launcher": "2.2.0",
     "karma-coverage": "1.1.2",
@@ -99,31 +99,37 @@
     "karma-mocha": "1.3.0",
     "karma-mocha-reporter": "2.2.5",
     "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "3.0.0",
-    "less": "3.0.4",
+    "karma-webpack": "3.0.5",
+    "less": "3.9.0",
     "less-loader": "4.1.0",
-    "lint-staged": "7.1.2",
-    "mini-css-extract-plugin": "0.4.0",
+    "lint-staged": "8.1.0",
+    "mini-css-extract-plugin": "0.5.0",
     "mocha": "5.2.0",
-    "pofile": "1.0.10",
-    "prettier": "1.12.1",
-    "raw-loader": "0.5.1",
-    "react-test-renderer": "16.4.0",
-    "remap-istanbul": "0.11.1",
-    "style-loader": "0.21.0",
-    "ts-loader": "4.3.0",
-    "tslint": "5.10.0",
-    "typescript": "2.8.3",
-    "url-loader": "1.0.1",
-    "webpack": "4.8.3",
-    "webpack-bundle-analyzer": "2.13.1",
-    "webpack-cli": "2.1.4",
-    "webpack-dev-server": "3.1.4"
+    "pofile": "1.0.11",
+    "prettier": "1.15.3",
+    "raw-loader": "1.0.0",
+    "react-test-renderer": "16.7.0",
+    "remap-istanbul": "0.13.0",
+    "style-loader": "0.23.1",
+    "ts-loader": "5.3.3",
+    "tslint": "5.12.1",
+    "typescript": "3.2.2",
+    "url-loader": "1.1.2",
+    "webpack": "4.28.4",
+    "webpack-bundle-analyzer": "3.0.3",
+    "webpack-cli": "3.2.1",
+    "webpack-dev-server": "3.1.14"
   },
   "lint-staged": {
-    "ignore": ["**/*.min.js", "**/*.min.css"],
+    "ignore": [
+      "**/*.min.js",
+      "**/*.min.css"
+    ],
     "linters": {
-      "*.{ts,tsx,js,json,css,less}": ["prettier --write", "git add"]
+      "*.{ts,tsx,js,json,css,less}": [
+        "prettier --write",
+        "git add"
+      ]
     }
   }
 }
