@@ -1,5 +1,4 @@
-import { Subject } from "rxjs/Subject";
-import { BehaviorSubject } from "rxjs/BehaviorSubject";
+import { Subject, BehaviorSubject } from "rxjs";
 
 export interface ICreateScreenshotCommand {
   isSmall: boolean;
